+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.2.0] - 2018-03-06
+
 ### Added
 
 - Hot reload option of renderer class
@@ -26,4 +28,5 @@
 - Runtime parameters
 
 [Unreleased]: https://github.com/aramvisser/navtastic/compare/v0.1.0...HEAD
-[0.1.0]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.0.1...v0.1.0
+[0.1.0]: https://github.com/aramvisser/navtastic/compare/v0.0.1...v0.1.0
+[0.2.0]: https://github.com/aramvisser/navtastic/compare/v0.1.0...v0.2.0